pyproject.toml
[project]
name = "reframeq-backend"
version = "0.1.0"
description = "ReframeQ API layer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "redis>=5.0.7",
  "celery>=5.4.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "httpx>=0.27.0",
  "ruff>=0.6.0"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*"]
